Factor out the concept of a buffer queue from that of a byte buffer queue.
