+++ src/stream/ByteQueue.h
@@ -16,193 +16,16 @@
  */
 
 // Defines a byte queue for hold bit and byte streams.
-//
-// Associated with the byte queue are locks.  This locking structure allows one
-// to lock "backpatch" addresses, making sure that the pages are not thrown away
-// until all locks have been released.  Locks will also block reads, until the
-// locks are unlocked.
-//
-// Note: Virtual addresses are used, start at index 0, and correspond to a
-// buffer index as if the queue keeps all pages (i.e. doesn't shrink) until the
-// queue is destructed. Therefore, if a byte is written at address N, to read
-// the value you must always use address N.
-//
-// It is assumed that jumping on reads and writes are valid. However, back jumps
-// are only safe if you lock the address before reading past that address.
-//
-// TODO(KarlSchimpf): Locking of reads/writes for threading has not yet been
-// addressed, and the current implementation is NOT thread safe.
-//
-// TODO(KarlSchimpf): Make reads blocked if after a frozen page.
-//
-// TODO(KarlSchimpf): bocking not implemented. Currently, reads fail if they try
-// to read from a locked page.
 
-#ifndef DECOMPRESSOR_SRC_STREAM_QUEUE_H
-#define DECOMPRESSOR_SRC_STREAM_QUEUE_H
+#ifndef DECOMPRESSOR_SRC_STREAM_BYTEQUEUE_H
+#define DECOMPRESSOR_SRC_STREAM_BYTEQUEUE_H
 
-#include "stream/Page.h"
+#include "stream/Queue.h"
 #include "stream/RawStream.h"
-
-#include <cstring>
-#include <memory>
-#include <queue>
-#include <vector>
 
 namespace wasm {
 
 namespace decode {
-
-template<class Base>
-class Queue {
-  Queue(const Queue &) = delete;
-  Queue &operator=(const Queue &) = delete;
-
-public:
-  Queue();
-
-  virtual ~Queue();
-
-  // Defines the maximum Peek size into the queue when reading. That
-  // is, The minimal number of bytes that the reader can back up without
-  // freezing an address. Defaults to 32.
-  void setMinPeekSize(size_t NewValue) {
-    MinPeekSize = NewValue * sizeof(Base);
-  }
-
-  // Value unknown (returning maximum possible size) until frozen. When
-  // frozen, returns the size of the buffer.
-  size_t currentSize() {
-    return EobFrozen
-        ? EobPage->getMaxAddress() : std::numeric_limits<ssize_t>::max();
-  }
-
-  // Returns the actual size of the buffer (i.e. only those with pages still
-  // in memory).
-  size_t actualSize() const {
-    return EobPage->getMaxAddress() - FirstPage->getMinAddress();
-  }
-
-  // Returns true if Address is locked.
-  bool isAddressLocked(size_t Address) const {
-    Page *P = getPage(Address);
-    if (P == nullptr)
-      return false;
-    return P->isLocked();
-  }
-
-  // Increments the lock count for Address by 1. Assumes lock is already locked
-  // (and hence defined).
-  void lock(size_t Address) { lockPage(getPage(Address)); }
-
-  // Decrements the lock count for Address by 1. Assumes lock was
-  // defined by previous (successful) calls to getReadLockedPointer()
-  // and getWriteLockedPointer().
-  void unlock(size_t Address) { unlockPage(getPage(Address)); }
-
-
-  // The following two methods allows one to lock the memory of the queue
-  // directly, and read/write directly into the buffer. Also locks the
-  // address, which must be released with a call to unlockAddress().
-  //
-  // WARNING: Never access beyond LockedSize elements after the call,
-  // and only while you haven't unlocked the address.  There is no
-  // guarantee that such pointer accesses are vaild, once it has been
-  // unlocked.
-
-  // Returns a pointer into the queue that can be read. Must unlock the address
-  // once reading has been completed.
-  //
-  // @param Address    The address within the queue to Access.
-  // @param WantedSize The number of elements requested to be read
-  //                   locked.  Note: Zero is allowed.
-  // @param LockedSize The actual number of elements locked down (may
-  //                   be smaller than wanted size, due to eob or
-  //                   internal paging).
-  // @result           Pointer to locked address, or nullptr if unable to read
-  //                   lock.
-  uint8_t *getReadLockedPointer(size_t Address, size_t WantedSize,
-                                size_t &LockedSize);
-
-  // Returns a pointer into the queue that can be written to. Must unlock the
-  // address once writing has been completed.
-  //
-  // @param Address    The address within the queue to Access.
-  // @param WantedSize The number of elements requested to bewrite
-  //                   locked.  Note: Zero is allowed, and just locks
-  //                   address.
-  // @param LockedSize The actual number of elements locked down (may
-  //                   be smaller than wanted size, due to eob or
-  //                   internal paging).
-  // @result           Pointer to locked address, or nullptr if unable to write
-  //                   lock.
-  uint8_t *getWriteLockedPointer(size_t Address, size_t WantedSize,
-                                 size_t &LockedSize);
-
-  // Freezes eob of the queue. Not valid to read/write past the eob, once set.
-  void freezeEob(size_t Address);
-
-  bool isEobFrozen() const { return EobFrozen; }
-
-protected:
-  // Minimum peek size to maintain. That is, the minimal number of
-  // bytes that the read can back up without freezing an address.
-  size_t MinPeekSize = 32 * sizeof(Base);
-  // True if end of queue buffer has been frozen.
-  bool EobFrozen = false;
-  // First page still in queue.
-  Page *FirstPage;
-  // Page at the current end of buffer.
-  Page *EobPage;
-  // Fast page lookup map (from page index)
-  using PageMapType = std::vector<Page*>;
-  PageMapType PageMap;
-  // Heap to keep track of pages locks, sorted by page index.
-  std::priority_queue<size_t, std::vector<size_t>,
-                      std::less<size_t>> LockedPages;
-
-  // Returns the page in the queue referred to Address, or nullptr if no
-  // such page is in the byte queue.
-  Page *getPage(size_t Address) const {
-    return getPageAt(Page::index(Address));
-  }
-
-  // Returns the page with the given PageIndex, or nullptr if no such
-  // page is in the byte queue.
-  Page *getPageAt(size_t PageIndex) const {
-    return (PageIndex >= PageMap.size()) ? nullptr : PageMap[PageIndex];
-  }
-
-  // Increments the lock count on the given page.
-  void lockPage(Page *P) {
-    P->lock();
-    LockedPages.emplace(P->Index);
-  }
-
-  // Decrements the lock count on the given page.
-  void unlockPage(Page *P) {
-    P->unlock();
-    // Remove smallest page indices from queue that no longer have locks.
-    while (!LockedPages.empty()) {
-      size_t PageIndex = LockedPages.top();
-      Page *P = getPageAt(PageIndex);
-      if (P && P->isLocked())
-        return;
-      LockedPages.pop();
-    }
-  }
-
-  // Dumps and deletes the first page.  Note: Dumping only occurs if a
-  // Writer is provided (see class WriteBackedByteQueue below).
-  virtual void dumpFirstPage();
-
-  // Dumps ununsed pages before the given address to recover memory.
-  void dumpPreviousPages(size_t Address);
-
-  // Fills buffer until we can read 1 or more bytes at the given address.
-  // Returns true if successful.
-  virtual bool readFill(size_t Address);
-};
 
 class ByteQueue : public Queue<uint8_t> {
   ByteQueue(const ByteQueue &) = delete;
@@ -213,17 +36,14 @@ public:
 
   ~ByteQueue() override {}
 
-  // Reads a contiguous range of elements into a buffer.
+  // Reads a contiguous range of bytes into a buffer.
   //
   // Note: A read request may not be fully met. This function only guarantees
   // to read 1 element from the queue, if eob hasn't been reached. This is
   // done to minimize blocking. When possible, it will try to meet the full
   // request.
   //
-  // TODO(KarlSchimpf): Add lock blocking if elements are not yet available.
-  // Rather, than blocking, the current implementation returns 0.
-  //
-  // @param Address The addres within the queue to read from. Automatically
+  // @param Address The address within the queue to read from. Automatically
   //                incremented during read.
   // @param Buffer  A pointer to a buffer to be filled (and contains at least
   //                Size elements).
@@ -232,16 +52,16 @@ public:
   //                reached. Valid return values are in [0..Size].
   size_t read(size_t &Address, uint8_t *Buffer, size_t Size=1);
 
-  // Writes a contiquous sequence of elements in the given buffer.
+  // Writes a contiquous sequence of bytes in the given buffer.
   //
   // Note: If Address is larger than queue size, zero's are automatically
   // inserted.
   //
-  // @param Address    The addres within the queue to write to. Automatically
-  //                   incremented during write.
-  // @param Buffer     A pointer to the buffer of elements to write.
-  // @param Size       The number of elements in the buffer to write.
-  // @result           True if successful (i.e. not beyond eob address).
+  // @param Address The address within the queue to write to. Automatically
+  //                incremented during write.
+  // @param Buffer  A pointer to the buffer of elements to write.
+  // @param Size    The number of elements in the buffer to write.
+  // @result        True if successful (i.e. not beyond eob address).
   bool write(size_t &Address, uint8_t *Buffer, size_t Size=1);
 
   // For debugging. Writes out sequence of bytes (on page associated with
@@ -291,4 +111,4 @@ private:
 
 } // end of namespace wasm
 
-#endif // DECOMPRESSOR_SRC_STREAM_QUEUE_H
+#endif // DECOMPRESSOR_SRC_STREAM_BYTEQUEUE_H
